initial unit tests accuracy.py, action.ExecutionState, action.Action and action.Move completed

+++ simulator/accuracy.py
@@ -2,5 +2,5 @@ from decimal import Decimal
 
 precision = Decimal("0.000")
 
-def quantize(float):
-	return Decimal.from_float(float).quantize(precision)
+def quantize(float_):
+	return Decimal.from_float(float_).quantize(precision)
